+++ ReleaseNotes.md
@@ -1,5 +1,5 @@
 # 2022.xx.xx.xxxxx
 
-* 
+* Fix some false negatives for `AvoidSwitchStatementsRule` (where the switch is on a `string` value)
 
 ⁋For previous releases, go here -- https://github.com/SteveGilham/Gendarme/blob/trunk/ReleaseNotes%20-%20Previously.md